Extracted measurement unit logic into its own module
In order to attach some more values to them (such as name, abbreviation),
I changed the units from enums to enum-like classes. Then in order to
minimise repeated code while maintaining type safety, I employed templates.

This might potentially be a bit inefficient in the future if we end up
having many many many unit types, but I don't suspect this to be the case.